[Refactor] STEP1 - WeatherDetailViewController로부터 화면 그리는 역할을 분리
화면을 그리는 역할은 View에서 수행하도록 DetailView를 생성해 WeatherDetailViewController에서 책임을 덜어낸다.

Fixes: https://camp.yagom-academy.kr/camp/65cc2b983173745319e0464c/projects/65cc31bc3173745319e0465e
